Respec Sequence->slice() to have starting index and count of items
The starting index must be inclusive to retrieve the first item. The ending index must be inclusive to retrieve the last last item. However, because both must be inclusive, there is no way to retrieve an empty list of items, and such a situation should be reachable, for certain situations. For this reason, dropping the ending index entirely, and just using a count of the items to retrieve solves the problem entirely, while reducing code complexity.